+++ src/drive_plausibility.cpp
@@ -1,6 +1,6 @@
 #include <Arduino.h>
 #include <imxrt.h>
-#include <main.h>
+#include "machine.h"
 //#include <fakenodes.h>
 
 int d_plaus(iCANflex &Car){
